+++ example_futures_commission_rate.py
@@ -47,7 +47,10 @@ logging.basicConfig(level=logging.DEBUG,
 api_key = ""
 api_secret = ""
 
+# To use this library you need a valid UNICORN Binance Suite License: https://medium.lucit.tech/87b0088124a8
 ubra = BinanceRestApiManager(api_key, api_secret, exchange="binance.com-futures")
 
 print(ubra.futures_commission_rate(symbol='BTCUSDT'))
+
+ubra.stop_manager()