+++ example_logging.py
@@ -45,7 +45,11 @@ logging.basicConfig(level=logging.DEBUG,
 
 api_key = "aaa"
 api_secret = "bbb"
+
+# To use this library you need a valid UNICORN Binance Suite License: https://medium.lucit.tech/87b0088124a8
 ubra = BinanceRestApiManager(api_key, api_secret)
 
 depth = ubra.get_order_book(symbol='BNBBTC')
 print(f"{depth}")
+
+ubra.stop_manager()